This is a re-working of the buildroot configs and setup
ChangeLog:

- Switch buildroot to 2022.02.4
- Add a CUSTOM_CONFIG option to runme.sh so you can easily
choose the buildroot config to use
- Change buildroot configs to use BR2_GLOBAL_PATCH_DIR rather
than separate patch directories for the kernel and u-boot. This
simplifies upgrades as you just need to make a new directory in
patches/package/version and add patches and they will be included
in the build
- Changed the patch structure as required by the above change
- Switched BUILDROOT to use BR2_ROOTFS_MERGED_USR which is the
new standard for Linux filesystems. This means that all / directories
are linked into /usr
- Changed the overlay directory to reflect the requirements of enabling
the above config option.
- Added kernel patches that never made it into mainline that enable
mipi-dsi support for the iMX6 SOCs

Signed-off-by: Jon Nettleton <jon@solid-run.com>